requirements.txt
asgiref==3.2.7
bson==0.5.8
dataclasses==0.6
Django==2.2.12
django-crispy-forms==1.9.1
django-mongoengine==0.4.1
django-mongoengine-forms==0.4.6
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
djongo==1.3.2
dnspython==1.16.0
mongoengine==0.20.0
Pillow==7.1.2
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
